src/requirements.txt
vosk
pyaudio
numpy
resemblyzer
scikit-learn
matplotlib
umap-learn
seaborn